API getting list useable accounts of an user for admin

+++ src/main/java/fpt/banking/system/service/AccountService.java
@@ -6,7 +6,7 @@ import fpt.banking.system.model.Account;
 
 public interface AccountService {
 
-	public List<Account> getAccounts(int userId);
+	public List<Account> getAccounts(long userId);
 	
 	public List<Account> getUseableAccounts(long userId);
